refactor(logging): 精简安卓调试日志工具实现
- 移除 SettingsStore 依赖及调试开关控制逻辑，简化日志输出函数
- 取消日志开关判断，所有调试日志均直接输出
- 修改日志调用处，去除条件判断以便调试信息始终打印
- 保持日志格式与 VueTagLogger 一致，方便浏览器端和安卓端统一查看
- 增加对推荐世界、加载世界等关键方法的日志输出，提升调试信息丰富度

+++ app/src/main/java/com/toonflow/game/data/Models.kt
@@ -32,7 +32,7 @@ data class RoleParameterCard(
   @SerializedName("equipment") val equipment: List<String> = emptyList(),
   @SerializedName("hp") val hp: Int = 100,
   @SerializedName("mp") val mp: Int = 0,
-  @SerializedName("money") val money: Int = 0,
+  @SerializedName("money") val money: Long = 0L,
   @SerializedName("other") val other: List<String> = emptyList(),
   @SerializedName("executing_task") val executingTask: JsonElement? = null,
   @SerializedName("cultivation_progress") val cultivationProgress: JsonElement? = null,
